Split econ-runs routes from scenarios.py into econ_runs.py
